Add comparison of null/nil safety and syntax
I researched these features because I was wondering if either language had a solution like Rust’s `Option`. Neither language does.

I decided the comparison result was “?” because D is better by default because more of its built-in types are non-nullable, but Nim can be made safer than D if you put in the work of redefining everything using the experimental `not nil` type annotation.